moved class types declarations in class folder
