+++ src/pages/Editor/Classes/UpdateClassModals/AddNewClassModal.tsx
@@ -2,7 +2,8 @@ import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { useParams } from 'react-router-dom';
 import { Input, InputField, TextArea } from 'components/Input';
-import { ErrorMessageInterface, AddNewClassFormProps } from 'types';
+import { ErrorMessageInterface } from 'types';
+import { AddNewClassFormProps } from 'pages/Editor/Classes/classTypes';
 import { createClass, CreateClassProps } from 'api/classes';
 import { FormErrorMessages, StyledErrorMessage } from 'components/ErrorMessage';
 import { Form } from 'components/Form';
