+++ src/pages/Editor/Classes/UpdateClassModals/EditClassModal.tsx
@@ -5,7 +5,8 @@ import React, { useEffect, useState } from 'react';
 import { useFieldArray, useForm } from 'react-hook-form';
 import styled from 'styled-components';
 import { Input, InputField, TextArea } from 'components/Input';
-import { ErrorMessageInterface, EditClassModalProps } from 'types';
+import { ErrorMessageInterface } from 'types';
+import { EditClassModalProps } from 'pages/Editor/Classes/classTypes';
 import { CreateClassProps, editClass } from 'api/classes';
 
 import { FormErrorMessages, StyledErrorMessage } from 'components/ErrorMessage';
